Conclusão das funçoes do estoque

+++ ECommerceAPI/ECommerceAPI/Repository/Impl/EstoqueRepositoryImpl.cs
@@ -1,6 +1,5 @@
 ﻿using ECommerceAPI.Models;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Query;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -11,26 +10,36 @@ namespace ECommerceAPI.Repository
     public class EstoqueRepositoryImpl : EstoqueRepository
     {
         private readonly ECommerceDbContext _context;
-
         public EstoqueRepositoryImpl(ECommerceDbContext context)
         {
             _context = context;
         }
-        public void Atualizar(Estoque estoque)
+
+        public bool Atualizar(Estoque estoque)
         {
-            if (!Exists(estoque))
+            if (Exists(estoque))
             {
-                _context.Update(estoque);
+                _context.Entry(estoque).State = EntityState.Modified;
                 _context.SaveChanges();
+                return true;
+            }
+            else
+            {
+                return false;
             }
         }
 
-        public void Cadastrar(Estoque estoque)
+        public bool Cadastrar(Estoque estoque)
         {
             if (!Exists(estoque))
             {
                 _context.Add(estoque);
                 _context.SaveChanges();
+                return true;
+            }
+            else
+            {
+                return false;
             }
         }
 
@@ -39,11 +48,18 @@ namespace ECommerceAPI.Repository
             return _context.Estoques.Include(e => e.Produto).ToList();
         }
 
-        public void Remover(int id)
+        public bool Remover(int id)
         {
             var estoque = SelecionarPorId(id);
+
+            if (estoque == null)
+            {
+                return false;
+            }
+
             _context.Estoques.Remove(estoque);
             _context.SaveChanges();
+            return true;
         }
 
         public Estoque SelecionarPorId(int id)
@@ -53,7 +69,7 @@ namespace ECommerceAPI.Repository
 
         private bool Exists(Estoque estoque)
         {
-            return _context.Estoques.Any(e => (e.Id == estoque.Id));
+            return _context.Estoques.Any(e => (e.Id == estoque.Id || e.IdProduto == estoque.IdProduto));
         }
     }
 }
